feat: 유저 관심종목 조회,생성,삭제 구현
Jira: S11P21A607-350

+++ backend/spring/src/main/java/com/example/mutualrisk/asset/repository/AssetRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AssetRepository extends JpaRepository<Asset, Long> {
+public interface AssetRepository extends JpaRepository<Asset, Integer> {
     @Query("SELECT a FROM Asset a WHERE a.name LIKE %:keyword% OR a.code LIKE %:keyword%")
     List<Asset> searchByKeyword(@Param("keyword") String keyword);
 }
